+++ GreenBankX/GreenBankX/AddTreePop.xaml.cs
@@ -30,7 +30,7 @@ namespace GreenBankX
             if (MHeight.Text != null && double.Parse(MHeight.Text) > 0 && Diameter.Text != null && double.Parse(Diameter.Text) > 0 && Application.Current.Properties["Counter"] != null && (int)Application.Current.Properties["Counter"] > -1 && DateMes.Date < DateTime.Now)
             {
                 counter = (int)Application.Current.Properties["Counter"];
-                ((List<Plot>)Application.Current.Properties["Plots"]).ElementAt(counter).AddTree(new Tree(float.Parse(Diameter.Text), float.Parse(MHeight.Text), ((List<Plot>)Application.Current.Properties["Plots"]).ElementAt(counter).getTrees().Count, DateMes.Date));
+                ((List<Plot>)Application.Current.Properties["Plots"]).ElementAt(counter).AddTree(new Tree(double.Parse(Diameter.Text), double.Parse(MHeight.Text), ((List<Plot>)Application.Current.Properties["Plots"]).ElementAt(counter).getTrees().Count, DateMes.Date));
                 Application.Current.Properties["Counter"] = -1;
                 MessagingCenter.Send<AddTreePop>(this, "Add");
                 await PopupNavigation.Instance.PopAsync();
