+++ GreenBankX/GreenBankX/MainPage.xaml.cs
@@ -26,6 +26,7 @@ namespace GreenBankX
         {
             await Navigation.PushAsync(new MenuPage());
         }
+        //loads data from .xls files. prices. data for plots is stored in Pricings.xls
         void LoadPriceFiles()
         {
             bool doesExist = File.Exists(DependencyService.Get<ISave>().GetFileName() + "/Pricings.xls");
@@ -53,6 +54,7 @@ namespace GreenBankX
             }
 
         }
+        //loads data from .xls files plots. data for plots is stored in Plots.xls
         void LoadPlotFiles()
         {
             bool doesExist = File.Exists(DependencyService.Get<ISave>().GetFileName() + "/Plots.xls");
@@ -91,7 +93,7 @@ namespace GreenBankX
 
             }
         }
-
+        //loads data from .xls files populates plots with trees. data for trees is stored in <PlotName>.xls
         void LoadTreeFiles()
         {
             for (int z = 0; z < ((List<Plot>)Application.Current.Properties["Plots"]).Count; z++) {
@@ -112,11 +114,11 @@ namespace GreenBankX
                         string ID = sheet.GetValueRowCol(1, 2).ToString();
                             SortedList<DateTime, (double, double)> History= new SortedList<DateTime, (double, double)>();
 
-                        Tree newTree = new Tree(float.Parse(sheet.GetValueRowCol(3, 2).ToString()), float.Parse(sheet.GetValueRowCol(3, 3).ToString()), int.Parse(sheet.GetValueRowCol(1, 2).ToString()),DateTime.Parse((sheet.GetValueRowCol(3, 1).ToString())));
+                        Tree newTree = new Tree(double.Parse(sheet.GetValueRowCol(3, 2).ToString()), double.Parse(sheet.GetValueRowCol(3, 3).ToString()), int.Parse(sheet.GetValueRowCol(1, 2).ToString()),DateTime.Parse((sheet.GetValueRowCol(3, 1).ToString())));
                         for (int y = 1; y < int.Parse(sheet.GetValueRowCol(1, 3).ToString()); y++)
                         {
-                                float.Parse(sheet.GetValueRowCol(3 + y, 2).ToString());
-                                newTree.AddToHistory(float.Parse(sheet.GetValueRowCol(3+y, 2).ToString()), float.Parse(sheet.GetValueRowCol(3+y, 3).ToString()), DateTime.Parse((sheet.GetValueRowCol(3+y, 1).ToString())));
+                                double.Parse(sheet.GetValueRowCol(3 + y, 2).ToString());
+                                newTree.AddToHistory(double.Parse(sheet.GetValueRowCol(3+y, 2).ToString()), double.Parse(sheet.GetValueRowCol(3+y, 3).ToString()), DateTime.Parse((sheet.GetValueRowCol(3+y, 1).ToString())));
                         }
                             ((List<Plot>)Application.Current.Properties["Plots"]).ElementAt(z).AddTree(newTree);
                     }
